added sorting for Student List table
