+++ src/components/admin/content/student/style.scss
