+++ src/index.tsx
@@ -2,7 +2,6 @@ import React from "react"
 import ReactDOM from "react-dom/client"
 import "./themes/global.scss"
 import App from "./App"
-import { loadUser } from "contexts/loginReducer"
 import { store, persistor } from "./utils/store"
 import { PersistGate } from "redux-persist/integration/react"
 import { Provider } from "react-redux"
